handle received number fields NumberFormatException

+++ bundles/org.openhab.binding.nikohomecontrol/src/main/java/org/openhab/binding/nikohomecontrol/internal/protocol/nhc2/NhcThermostat2.java
@@ -54,7 +54,7 @@ public class NhcThermostat2 extends NhcThermostat {
     }
 
     @Override
-    public void executeOverrule(int overrule, int overruletime) {
+    public void executeOverrule(long overrule, long overruletime) {
         logger.debug("execute thermostat overrule {} during {} min for {}", overrule, overruletime, id);
 
         nhcComm.executeThermostat(id, overrule, overruletime);
